Make VerificationError available in the verification API
