+++ sigstore/verify/verifier.py
@@ -48,7 +48,9 @@ from sigstore._internal.rekor.client import RekorClient
 from sigstore._internal.set import InvalidSETError, verify_set
 from sigstore._internal.tuf import TrustUpdater
 from sigstore._utils import B64Str, HexStr
+from sigstore.verify.models import CertificateVerificationFailure
 from sigstore.verify.models import InvalidRekorEntry as InvalidRekorEntryError
+from sigstore.verify.models import LogEntryMissing
 from sigstore.verify.models import RekorEntryMissing as RekorEntryMissingError
 from sigstore.verify.models import (
     VerificationFailure,
@@ -59,42 +61,6 @@ from sigstore.verify.models import (
 from sigstore.verify.policy import VerificationPolicy
 
 logger = logging.getLogger(__name__)
-
-
-class LogEntryMissing(VerificationFailure):
-    """
-    A specialization of `VerificationFailure` for transparency log lookup failures,
-    with additional lookup context.
-    """
-
-    reason: str = (
-        "The transparency log has no entry for the given verification materials"
-    )
-
-    signature: B64Str
-    """
-    The signature present during lookup failure, encoded with base64.
-    """
-
-    artifact_hash: HexStr
-    """
-    The artifact hash present during lookup failure, encoded as a hex string.
-    """
-
-
-class CertificateVerificationFailure(VerificationFailure):
-    """
-    A specialization of `VerificationFailure` for certificate signature
-    verification failures, with additional exception context.
-    """
-
-    reason: str = "Failed to verify signing certificate"
-    exception: Exception
-
-    class Config:
-        # Needed for the `exception` field above, since exceptions are
-        # not trivially serializable.
-        arbitrary_types_allowed = True
 
 
 class Verifier:
